tast/webrtc: fix/simplify DecodeAccelUsedJPEG
This CL simplifies DecodeAccelUsedJPEG to not use the code in
loopback.html code, instead adding a tiny get_user_media.html
and running the call to the promisified getUserMedia() in
video/video.go (that will be renamed later to get_user_media.go
to simplify reviews).

The function checkError() in video.go is not used anymore for
this test, so it's migrated to its only caller in perf.go.

This test is notoriously flaky in small devices due to the
large overhead of setting an RTCPeerConnection, with at least
an encoder and a decoder; with this simplification the flakes
are expected to be reduced.

BUG=chromium:1018590

TEST=tast -verbose run papaya webrtc.*AccelUsed* on kohaku

Change-Id: I3f20ac2919c1ed39257936632c99e262c8f3d4be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1879965
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>