tast/graphics: Add DRMAtomic test
WIP WIP WIP WIP

This CL adds a new test DRMAtomic, that runs a new function
IsDRMAtomicSupported() to figure out if the Atomic DRM is
supported or not by the (a) driver in the system.

It then compares that with the drm_atomic capability of the
board (which comes from the IUSE flags). Note that adding
drm_atomic is a manual process and we have boards (e.g. hatch)
that do support atomic but haven't got it enabled, which
prevents e.g. video overlays and low latency canvas.

BUG=chromium:1019869,chromium:1019386

TEST=tast run HERE_IP graphics.DRMAtomic on a hatch, a babymega and a scarlet

Change-Id: Ie5af4a4781b91c27e8cff320c63ff5d0c4df2234