tast/graphics: Add graphics.DRM.verify_drm_atomic
This CL adds a new test to DRM that verifies if the capability
drm_atomic presence or absence correlates with support on the
DRM API. drm_atomic comes from the IUSE flags on the chipset;
note that adding such flag is a manual process and we have
boards (e.g. hatch) that do support atomic but haven't got it
enabled.

A new file graphics/drm/utils.go is added with some support
code that uses ioctl() and iow(), present in tast/local/input;
This CL makes those symbols public.

BUG=chromium:1019869,chromium:1019386

TEST=tast run HERE_IP graphics.DRM.* on a hatch, a babymega and a scarlet

Change-Id: Ie5af4a4781b91c27e8cff320c63ff5d0c4df2234