WIP: tast: fix a deadlock in dbusutil.SignalWatcher
When godbus receives a signal, godbus.defaultSignalHandler holds a RLock
of the channel slice and writes to subscribed channels. In RemoveSignal,
godbus.defaultSignalHandler acquires a full Lock of the slice to remove
the channel. However, if the RemoveSignal caller stops reading the
channel, there can be a deadlock. To unravel this, we should start the
Signals consumer before RemoveSignal.

TODO: create a bug for this.
TODO: the test is awkward because it need dbus to run, but we don't have
one in chroot.

BUG=None
TEST=None
Change-Id: I0501064a224a77aa61fb063e955f735be19c3b29