tast: fix a deadlock in dbusutil.SignalWatcher
When godbus receives a signal, godbus.defaultSignalHandler holds a RLock
of the channel slice and writes to subscribed channels. In RemoveSignal,
godbus.defaultSignalHandler acquires a full Lock of the slice to remove
the channel. However, if the RemoveSignal caller stops reading the
channel, there can be a deadlock. To unravel this, we should start the
Signals consumer before RemoveSignal.

BUG=chromium:1020058
TEST=tast run $DUT dbus.SignalWatcherClose

Change-Id: I0501064a224a77aa61fb063e955f735be19c3b29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1890573
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>