tast-tests: arc.WindowManagerCUJ flaky
This CL tries to fix the flakyness in arc.WindowManagerCUJ.
From the error logs it seems that it is failing on some boards
because it is trying to use UIAutomator while the activity is not ready.

Changing act.WaitForResumed() for ash.WaitForVisible() should fix this,
since WaitForResumed() does not guarantee that the activity is ready.

BUG=None
TEST=tast run $DUT arc.WindowManagerCUJ
     run it multiple times. passed the test all times.

Change-Id: Ia99b9c51332f5fae6be3012d11546ce0aae49e48