tast: Fix the existence of install_attributes files in tests
The presence of install_attributes is part of the system, and is used
by session_manager when determining if the device is enrolled.

The tests need to have the presence of the file according to the scenario
being tested. For now this change only creates the file with default
data if needed. For future we might want to set the actual data in
the file depending on whether we want the device to be enrolled or not.

BUG=chromium:1011375
TEST=tast run <ip>

Change-Id: I32ea77d98d66546f902a4a3f7f178e3c2ef39c44