tast/webrtc: extend RTCPeerConnectionAccelUsed to allow codecs
This CL extends the said test to have parameters, namely the profile
name (vp8, vp9 etc), and pipes that param to loopback.html, where
munge_sdp.js is used to configure the codec profile for the test.
The function in loopback.html is changed to return a Promise.

This CL does not add new tests, just prepares the ground for the
next one where H264 and VP9 will be added.

BUG=chromium:1017374
TEST=tast run -verbose papaya, webrtc.RTCL*AccelUsed* on babymega, kohaku

Change-Id: Ib60f34b5992460cd9329bdd52e37049674ce6099