video.DecodeEncodeAccelPerf: Increase timeout of decode encode perf tests.
This CL increases the timeout of the decode encode performance test. On
successful runs some devices take 1m45s for the test to complete. This is much
too close to the configured default timeout of 2m, as some time is reserved for
test cleanup and the test is aborted after running for 1m46s.

The timeout is increased from 2 minutes to 5 minutes to leave some margin to
prevent future flakiness.

TEST=video.DecodeEncodeAccelPerf on nocturne

BUG=chromium:1020411

Change-Id: Iaa9c5a4ec8bb7d7263dcd0ea2d1f1730faacebed