video.DecodeAccelPerf: Remove fallback on old performance report file name.
In crrev.com/c/1778104 the name of the video decoder performance tests in the
video_decode_accelerator_perf_tests binary was changed, which affected the name
of the generated performance test report.

To make sure that the Tast tests would work fine before and after the next
Chrome uprev a fallback was introduced to check both the old and new performance
test report files. This CL removes this fallback as it's no longer necessary.

TEST=video.DecodeAccelPerf.h264_1080p_30fps on nocturne

BUG=None

Change-Id: I50e44741809775836ea62b0e1575a836213edb25