tast/video: do not verify histograms in PlayDecodeAccel*
These tests verify that the error histogram has not been updated
after the test logic has run.  This is flaky for a number of
reasons, the most important one is that media/ can try a
number of potentially failing initializations before actually
finding a working one, and the failed ones are reflected in
the said histograms, making the test fail (*). This CL removes
the histogram checking, relying only on chrome://media-internals.

This has become somehow more common after R80-12636.0.0.

(*) e.g. on grunt I see media/ tries to start up V4L2 (fails),
then VAAPI (works). Overall result: acceleration is used, but
the histograms reflect a failure, and that makes the whole test
fail, which is incorrect.

BUG=chromium:1020560
TEST=tast run -verbose papaya, video.PlayDecodeAccel* on babymega, kohaku (hatch), kasumi (grunt)

Change-Id: I289875ca94c1a477e57c9b20681373431abc537f