tast-tests: New test newtork.HostapHwsim
This test sets up the harness for running the hostap project's hwsim
tests. These tests run in a VM, using virtualized WiFi clients and APs.
They intend to test the behaviors of the hostapd (AP) and wpa_supplicant
(client) daemons.

This is still a bit of WIP:
 * didn't finish the SoftwareDeps
 * runs only a very minimal test list (to be expanded)
 * is this really 'mainline' targeted? a full test list can take many
   minutes (possibly hours)
 * probably more

Stdout/stderr are logged in realtime to the Tast runner stdout, but the
full logs (including hostapd / wpa_supplicant logs) are stashed in your
SDK chroot at:
  /tmp/tast/results/${DATE}-${TIME}/tests/network.HostapHwsim/

BUG=chromium:1013471
TEST=run it on a Betty VM

Cq-Depend: chromium:1865763
Change-Id: Ic645ec83298692616f4b976f9a888c5c689a0970