tast/webrtc: extend RTCPeerConnectionAccelUsed to try vp9, h264
This CL extends the said test to try vp9 and/or h264 where
available. h264 has been supported by webrtc for a long time,
as it is mandatory, and vp9 has received plenty of attention
in the last years, particularly for screen/window/tab sharing,
so it's fitting to test it.

BUG=chromium:1017374
TEST=tast run -verbose papaya, webrtc.RTC*AccelUsed* on kohaku

Change-Id: Ic163c0f78fca11e0c9d5c5c4c100074077c3a34c