tast-tests: Fix bug where presentation time means were skewed.
The last bucket has a max of INT_MAX and a min of 200. If we had
any sample in here the mean will be calculated as a very large
value. For the mean calculation, use the min for the max in the
last bucket. Modifies ui.OverviewScrollPerf to run multiple times.

BUG=chromium:1020751
TEST=manual

Change-Id: Ic5b8bb945f6d958aae504d662ac0460230af30d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1897001
Tested-by: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>