tast/webrtc: remove histogram checking in DecodePerf
This CL removes the code in peer_connection_perf.go to avoid relying
on histogram updates, which turned out to be flaky in the past. This
is used from webrtc.DecodePerf Tast test.

Removing the histogram check allows to align this file with its
sibling peer_connection.go, used in RTCPeerConnectionAccelUsed.*

BUG=chromium:1017374
TEST=tast run -verbose papaya, webrtc.DecodePerf* on kohaku

Change-Id: Ia58e6fe862839a342f9f2bae321a28177e2034c0