Update patch set 1
Patch Set 1:

(11 comments)

Looks nice besides language nits, but if possible we should try to get rid of the locks and atomic setters/getters. What data actually needs to be shared by the server thread and the test thread ? If this is just signaling for stopping or starting the tests we should use channels instead. For instance instead of using a "stopped" variable to stop the server thread from the test thread, we should use a "cancellation" chan struct{} that is closed from the test side and which allows to exit the server loop without an explicit shared variable. See https://blog.golang.org/pipelines for an example.

Patch-set: 1
Reviewer: Gerrit User 1238096 <1238096@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>