video.ARCDecode: Update codec2 e2e decoder tests
This change updates the tests to use the c2_e2e_test.apk from the pi-arc
repository. This allows rendering to a surface, which provides more
accurate max FPS measurements as well as the ability to measure power,
cpu and dropped frames while rendering a video to the display.

BUG=b:144681449, b:143584325, b:142948390
TEST=tast run DUT video.ARCDecode*

Change-Id: Ia435bf94b164df1bd53db50d2921f3c381ce819b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1928476
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>