tast-tests: Make platform.Mosys parametrized.
platform.Mosys tests multiple commands. Some of them work reliably on all
boards, whereas a few have been consistently failing. We are
parameterizing the test to be able to promote some subtests to CQ while
we are investigating the failures.

Consistently passing commands were put in the test case with empty name.
Commands which are failing on some boards were put in separate cases.

BUG=chromium:1022222
TEST=tast -verbose run scarlet-dru platform.Mosys*

Change-Id: I923acd75be147c83ffd8939498181b4ad2f274d4