tast-tests: arc.WindowState.* flaky
This CL tries to fix the flakyness in arc.WindowState.* tests. After
setting the window state on the ash side, the tests verify the ASH and
Android window state. Due to the asynchronous architecture, there could
be a small delay from setting the window state on the ash side and
propagating it to the Android side. This CL polls for 10 seconds, for
Android to react to the window state.

BUG=None
TEST=Ran arc.WindowState.* 10 times on eve without a failure.
     (chroot)$ for i in {1..10};do  echo $i; tast run $DUT arc.WindowState.*; done
Change-Id: I23c5bf36acb98940b18c724eb349a968c135a9ce