tast/graphics: parameterize WebRTCVideoPlaybackDelay
This CL extends graphics.WebRTCVideoPlaybackDelay to accept a
parameter representing the codec profile to be used. Currently
this is always VP8 (by default), so it's the one chosen (later
CLs will add VP9 and H264). munge_sdp.js is roped in to allow
for codec specification in the offer negotiation.

No extra functionality is intended, just the parameterization.

BUG=chromium:1021542
TEST=tast -verbose run yahhoy graphics.WebRTCVideoPlaybackDelay.*

Change-Id: I7165ba7c48be4f3720c33c45f1a01fa2a1353326
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1899573
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>