tast-tests: fix getServicePaths() in shill/manager.go
Currently, m.getSerivcePaths() calls m.props.GetObjectPaths() before
creating a property watcher. That may encounter a case that a property
change happen just right after GetObjectPaths() and before
CreateWatcher(). If the property to watch is not changed afterward,
then pw.WaitAll() is blocked till timeout. To fix the issue, we follow
the logic in GetWifiInterface(): create a signal watcher before getting
a property.

Then the CL renames getSerivcePaths() to getObjectPaths() as
m.GetDevices() and m.GetProfiles() should also apply the logic to handle
the case that shill just restarts with empty Devices or Profiles.

BUG=chromium:1022271
TEST=tast run $DUT network.*

Change-Id: Iad40132463326ec46856e23556b31a6d9666ca1c