tast-test: arc.WindowManagerCUJ does not use private API
In order to run this test on Android R, the ArcWMTest apk
should not call any Android private API.

This CL replaces the data that was provided by Android private APIs with
ones from Ash, making the test run on P and R (as well as N).
In particular it replaces the caption data that was returned by the APK,
with Ash APIs with have similar functionality.

Additionally, by using the new Ash API, it makes the "Lights In/out"
less flaky by polling until the frame condition is met.

TEST=tast run $DUT arc.WindowManagerCUJ
BUG=b:143908501
Change-Id: I0a36d19175f02ca3202d223e3dbc2307c62d4518