tast-test: improve reliability in arc.WindowManagerCUJ
This CL improves the arc.WindowManagerCUJ test by adding "readyness"
chess both at the Android and Ash sides.

This is because this tests is failing on some devices, and from the
screenshots it seems that the test is trying to check the window state,
but the activity seems to be not ready.

BUG=None
TEST=tast run $DUT arc.WindowManagerCUJ locally 5*times. Worked Ok.

Change-Id: Iaee8a6c3e7295ac9e602df6bb7808e55d2cb339f