tast-test: improve reliability in arc.WindowManagerCUJ
This CL improves the arc.WindowManagerCUJ test by adding "readyness"
chess both at the Android and Ash sides.

This is because this tests is failing on some devices, and from the
screenshots it seems that the test is trying to check the window state,
but the activity seems to be not ready.

BUG=None
TEST=tast run $DUT arc.WindowManagerCUJ locally 5*times. Worked Ok.

Change-Id: Iaee8a6c3e7295ac9e602df6bb7808e55d2cb339f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1904640
Tested-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Prameet Shah <phshah@chromium.org>
Reviewed-by: Cici Ruan <cuicuiruan@google.com>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Stefan Kuhne <skuhne@chromium.org>
Auto-Submit: Ricardo Quesada <ricardoq@chromium.org>