tast-tests: add CPU and Power consumption to crostini.VimCompile
CPU and Power consumption data will be reported along with compile
time. MeasureCPUStart, MeasureCPUEnd functions are added to cpu.go
to get cpu usage when we do not know the duration in prior.
Ex: When we want to collect CPU usage data on a device during vim
compilation, we won't know how much time each DUT takes to compile
vim. In such cases we can use these functions.
Power consumption data is collected using channels and goroutines.
Because we have to specify the duration to use
"dump_intel_rapl_consumption --interval_ms", we couldn't make
use of it.

BUG=chromium:1012396
TEST=tast run $IP_ADDR crostini.VimCompile

Change-Id: I0aa6aa85ab4f82afcae91fd7dc5b804525a0c997
Signed-off-by: Sushma Venkatesh Reddy <sushma.venkatesh.reddy@intel.com>