tast.video.DecodeAccelSanity: Add additional log messages to tests.
The video.DecodeAccelSanity Tast tests runs the video_decode_accelerator_tests
binary with an invalid video stream, to check whether the video decoder can
handle invalid video streams without crashing.

As the video stream is invalid, the tests themselves are expected to fail. This
can be confusing for people looking at the test logs, so this CL adds additional
log messages to the tests to clarify its working.

TEST=video.video.DecodeAccelSanity.vp9_1 on nocturne

BUG=chromium:1022991

Change-Id: Ic796f497eae23b96dec2cf8847e1a8c1dec87f7b