DRAFT tast-test: soft-clear tpm utils + cryptohome tpm live test
This CL provides utils for soft-clearing the TPM and demonstrates the utils by
porting the autotest server test platform_CryptohomeTpmLiveTestServer to Tast
as a local test.

(DO NOT SUBMIT)
This CL is a draft and not correctly formatted yet (e.g., with gofmt). The way
CryptohomeTpmLiveTests takes TPM ownership and waits for TPM to be ready is also
subjected to be replaced with hwsec utils in other pending CLs like
chromium:1826587.

Cq-Depend: chromium:1928008
Change-Id: I68115ff833a2276cb33ee9bc2ca408e4be756e8b