tast-tests: Reduce flakiness for WorkspaceInsets test on arc.CompanionLibrary
In ChromeOS companion library, the workspace insets function can
return the insets from Framework side. For the test, it comparing
the result with the display info which including the workspace
insets info from Chrome side. The display info need through many
times numeric conversion, which may cause some floating numerical
errors.

In this CL, the test check the value and make sure the test would
not failure when error occurred.

BUG=b:142438505
BUG=b:141978622
TEST=tast -verbose run my_crbook arc.CompanionLibrary

Change-Id: I88cf282950d2308087a3134c020495d14561b43d