tast-test: Launch apps through Launcher
Created tests for launching apps of various types through the launcher.
Since it is testing launcher functionality, it does not simply call
app.Launch(). Currently, launching a crostini app pops up a message to
install crostini even is it is already installed. This issue will be
looked into in another CL.

Also adds a ui/launcher package for controlling the launcher directly
through the UI.

BUG=chromium:982527
TEST=tast run <eve> ui.LauncherSearch*

Depends-On: chromium-review/1849808
Change-Id: I16ed3cd2315d5b94b839fe325ed0d02fb9e57bbd