graphics: glbench a more compact failure reason
Currently we are putting the output image as part of the reason. This
md5 hash is not human readable and makes it hard to read from stainless.
This CL makes it only show the failed tests. +sorting the list.

BUG=chromium:1022115
TEST=tast run ${fizz} graphics.GLBench.crostini

Change-Id: I1eb784252c6cc01346f85ab2ed09e66c281e4d1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1915082
Tested-by: Po-Hsien Wang <pwang@chromium.org>
Auto-Submit: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>