Update patch set 5
Patch Set 5:

Hi Hidehiko, could you take a look at this?

Per the earlier discussion - Justin and I talked about this offline, and came to the conclusion that this is a worthwhile test now that telem can be invoked to test the sandboxed cros_healthd process. The actual value of sku_number should be checked in unit tests, and this is an integration test that verifies the SECCOMP filter for the code path taken by cros_healthd in response to a mojo call for CachedVpd. As long as we get back a non-empty string, cros_healthd didn't crash attempting to retrieve the info!

Patch-set: 5
Reviewer: Gerrit User 1302456 <1302456@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>