Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > I didn't go through this cl in much detail but I don't see the need if there's one object implementing this interface. Why not just use the rule struct itself everywhere?
> 
> Actually there are more than one struct implementing the interface, but it was not immediately obvious to me as well and I initially thought there were just one impl.
> 
> If you look further down in the file you ll find specialized test rules that embed the base struct and selectively override some functions of the interface.

This CL (and others in the chain) are still WIP. I understand that there are differences in the language idioms and I do still plan to address that. This initial upload was more or less just a strict translation for you all to see. Perhaps I should have been more clear by adding you to CC instead of R. Anyway, I have a bit of time now and I'm almost done cleaning everything up at this point so I'll leave it as is and remove WIP when I'm ready for review.

Patch-set: 4