Update patch set 2
Patch Set 2: Code-Review+1

chats are going on elsewhere, but I'm fine with this.

As noted there, it would be nice to have a little better notes on the bug.

It would be even better if we had "hardware dependency" support, so we could disable this only for certain boards.

Patch-set: 2
Reviewer: Gerrit User 1132414 <1132414@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1