tast.video.*Perf: Check whether process being measured runs for entire duration.
The cpu.MeasureProcessUsage function is used to measure cpu usage and power
consumption while running one or more processes. Currently there is no check
however to make sure the processes run for the entire measurement duration.

This CL makes sure that the test will fail if the process being measured exited
before the end of the measurement interval.

TEST=video.DecodeAccelPerf.h264_1080p_30fps on hana
     camera.DecodeAccelJPEGPerf on hana

BUG=chromium:1025114

Change-Id: Id1eb88fc2c17618754ff08bbce604ffbfaac84a8