Update patch set 1
Patch Set 1:

Hi,

Recently I started working on a "CPU power comsumption" refactor:
https://crrev.com/c/1917452

Instead of calling the "dump_intel_ralp" command, it parses the RAPL files.
The benefit, is that parsing the files is not blocking, and the "dump_intel_rapl" binary is blocking and you have to kind-of-know the duration value before running the command.

The CL is still in review, but it would be great if your CL is based on mine.

Patch-set: 1
Reviewer: Gerrit User 1202079 <1202079@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>