vm.StartCrosvm: stabilize test and remove use of imageloader
The vast majority of failures with vm.StartCrosvm have been either due to
timeouts (we don't boot as fast as we used to) or because of imageloader
failing to download the cros-termina component.

Update timeouts where appropriate to a reasonably conservative value. Switch
the test to use external data files, which are quite reliable, rather than
having imageloader on the DUT download the VM images. Remove the now unused
imageloader code in the vm package as well.

BUG=chromium:1025374
TEST=vm.StartCrosvm

Change-Id: Id58eaeb611ad0bfcc56a05c26b2e52c86f8572b0