tast-test: moves power.go from crosint to public
Moves the power.go[1] files that was in crosint to a public folder,
renaming the file to "cpucooldown.go" since that is the only functionaly
that it is left.

The power consumption functionality was moved to the rapl.go file (see
previous CL in the chain).

[1]: http://cs/chromeos_internal/src/platform/tast-tests-private/src/chromiumos/tast/local/bundles/crosint/arc/power/power.go

TEST=tast run $DUT --buildbundle=crosint \
          arc.RuntimePerf.sniper3d arc.AuthPerf.unmanaged
BUG=chromium:1022203

Change-Id: I9c7d89ba3ec5e90c58acbcc0922163c8da739dc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1920148
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Tested-by: Ricardo Quesada <ricardoq@chromium.org>
Auto-Submit: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>