wilco.ECEvent: Update logic to poll dev node for events
The EC event file only returns one event per read based on the
implementation in the kernel driver. Previously only the first event in
the queue would be read causing flakiness.

It is also possible for previous EC events to be buffered in the event
queue causing the test to always pass. To counteract this, drain the
queue before the test starts.

BUG=b:134059465
TEST=On dut, plug in and unplug AC power from the device a few times
(This generates EC events).
tast run ${DUT_IP} wilco.ECEvent (pass) - Power events are captured but
ignored.

Change-Id: I5380b45ebc9dce0eccc41b3f8b8c64f91e9a65d8