Add support of panel orientation and touch screen rotation
Sometimes the display panel can be physically rotated in a way
that the input.TouchscreenEventWriter will emit touch events
unexpectedly. ui.OverviewScrollPerf is an example of that;
it wants a horizontal swipe but this is interpret as a vertical
swipe on such device.

This CL relies on Chromium CL (crrev.com/c/1922925) and consists
of three parts:
- allow Display struct to keep the orientation of the display
  panel.
- allow input.TouchscreenEventWriter to support rotated UI.
- fix ui.OverviewScrollPerf to use those new features.

BUG=chromium:1022614
TEST=fast_build.sh -T; tast run ui.OverviewScrollPerf on krane

Change-Id: I4470f9ad4d5d4e8aad6824b37b336da9d238d1c9