Add support of panel orientation and touch screen rotation
Sometimes the display panel can be physically rotated in a way
that the input.TouchscreenEventWriter will emit touch events
unexpectedly. ui.OverviewScrollPerf is an example of that;
it wants a horizontal swipe but this is interpret as a vertical
swipe on such device.

This CL consists of three parts:
- add a new function to obtain the screen's rotation which
  can check the panel orientation too.
- allow input.TouchscreenEventWriter to support rotated UI.
- fix ui.OverviewScrollPerf to use those new features.

BUG=chromium:1022614
TEST=fast_build.sh -T; tast run ui.OverviewScrollPerf on krane

Change-Id: I4470f9ad4d5d4e8aad6824b37b336da9d238d1c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1923587
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>