tast: Add setup function for crash tests that require developer image behavior
Add setup function for crash tests that stashes existing crashes for the
duration of the test but does not alter standard developer image
behavior.

This function will execute the same steps as SetUpCrashTest()
excluding the creation of the /run/crash_reporter/crash-test-in-progress file.
This is necessary for tests that verify crash behavior that only occurs
on developer images (e.g. device coredump files being added to the crash
directory).

BUG=b:131183287
TEST=tast -verbose run dut platform.DevCoredump
Verify that the test passes and that the contents of /var/spool/crash
are the same before and after running the test.

Change-Id: I9e59dc7a397bf719873ff0bd2872041cd4c9c274