Fix arc.Booted() for ARCVM
ARCVM boot can take so long that the old connection is still around in the proxy. This connection will get deleted.
So waiting for 'already connected' will succeed, and we'll wait-for-device on a connection that will be deleted,
causing us to time out on the wait for device.

This change keeps us polling the adb connect until we see a properly connected device from adb devices. It is then
safe to wait-for-device, because it is the one we care about.

BUG=b:144814978

TEST=on coral astronaut, tast arc.Boot, doesn't time out on wait-for-device any more.

Change-Id: I7143a479a749dec2593a2245c251484c2fb995a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1925939
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Tested-by: Charles Dueck <cwd@chromium.org>
Commit-Queue: Charles Dueck <cwd@chromium.org>