Add perf related logic in CCAUIPerf and cca.go
To measure the performance for basic usage of CCA, we should add some
perf related logic in order to set the connection between CCA background
and Tast test.

Since we already have CCAUIPreviewPerf, we changed it to CCAUIPerf and
include measuremnet for other UI operations.

The corresponding CL on Chrome side is:
https://crrev.com/c/1880995
We should land that CL (Chrome) first before landing this CL (Tast).

BUG=b:141518806
TEST=tast run [DUT] camera.CCAUIPerf

Change-Id: Ic47bc7915d560ec83b1699103715473b2b1cfb57