Update patch set 1
Patch Set 1:

On which kernel are you testing veyron_minnie? 3.14 or 4.19? This *feels* like something we should be able to fix in the driver instead of the test (IIRC from 3.14, one of the 2 devices is a dummy device of sorts, and it probably shouldn't be involved at all), but I'm not sure.

Patch-set: 1
Reviewer: Gerrit User 1132414 <1132414@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>