network: iw: add option to dump cmd output for debugging
network.WifiCaps test has flaky failures (e.g. no 2.4GHz support), but
it can not be reproduced even on the same DUT in lab. Add an option to
dump the command outputs so that we can gather more info when it fails.

BUG=chromium:1026896
TEST=tast run $DUT network.WifiCaps network.IWScan
There should be iw command dump in WifiCaps test but not in IWScan.

Change-Id: I4dfa4ede613bb040d1e72ba3a38850b30ddafcf5