Update patch set 2
Patch Set 2:

> Just went through the CL, it is just like what I did in crrev.com/c/1901166 . It is quite verbose but it seems the only good way when we don't have a VLOG logging.

For this case, as we already have crrev.com/c/1928093 solving a bug, maybe we can first see how it goes and wait for VLOG.

> Wait. If you just need to vlog on specific method, for instance ListPhys(), how about just adding isDebug param on it? So we don't need to add Runner wrapper everywhere?

As it's the second parsing bug in iw, I kind of want it more general in case of future need. Or maybe I worry too much.

Patch-set: 2