Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> In general using named return values and modifying them in deferred calls in this way makes the code pretty hard to reason about -- errors returned from MeasureUsage or even from other cleanup calls might be dropped. I'm not sure if this is intentional behavior.
> 
> However, this was already done in the existing code, so that's not for this CL.

Thanks for reviewing! I was actually suggested to do it like this in an earlier CL. There is currently only one test that runs multiple processes at the same time, and in the very very rare case that both processes fail somehow, it seems acceptable to only return the last error (all errors are logged anyway).

I guess the only alternative is keeping all the commands in a list, and waiting for them at the end of the function manually. But this would complicate the code (e.g. on early-out).

Patch-set: 2