Update patch set 4
Patch Set 4:

We could change WaitXXX to not update the snapshot and not returning the diffs. We make r.Histogram() the only way to return diffs and update the snapshot there.

I think it is better than the current proposal. The downside is that user would have to explicitly call r.Histogram() to get the diffs. More typing but think the code is actually cleaner.

WDYT?

Patch-set: 4
Reviewer: Gerrit User 1000101 <1000101@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>