tast/video: unbreak video.PlayDecodeAccel*/PlaybackPerf*
crrev.com/c/1904341 changed the name of Chrome's media
internals category |is_platform_video_decoder| to
|kIsPlatformVideoDecoder| by surprise, breaking all the
mentioned tests. It also changed the name of a category
and the URL (from the site, to the element url), for
reasons.

This CL fixes that by supporting both the old and the
new names, and renaming the element that we look for
(s/player-frame/player-name/).

BUG=chromium:1028275,chromium:1023900,chromium:794255
TEST=tast -verbose run papaya video.PlayDecodeAccelUsed.* on kohaku

Change-Id: Ib06dccc6f694449c207667dd092ee71d6136ba3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1933560
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>