Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Any chance this change is related to failures I'm getting on a kernel change?
> 
> 11/25 12:58:01.602 INFO |             utils:0287| [stdout] platform.DLCService                  [ FAIL ] Failed to installexit status 70
> 
> See https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1931443
> 
> for failures.

Hi Grant,

Those errors related to DLCService are not because of any of your changes.

It's due to the fact that I've added this new change that does a periodic check.
https://cs.corp.google.com/chromeos_public/src/platform2/dlcservice/dlc_service.cc?l=389

It happens that you ran into a case where:
1) DLCService starts an install
2) Update Engine performs the DLC install
3) DLCService period check kicks in
3.5) During DLCService's period check that asks for Update Engine status synchronously, Update Engine sends a install completion signal, but will run after the Periodic check.
4) Periodic check thinks Update Engine is in a bad state. (Hence the tast failure)

I'll fix this soon, since this is a race condition on how the Status of DLC installs are handled.

Patch-set: 6