tast-tests: Run Crostini stretch download tests after buster
Change the precondition name so the buster download tests run sooner, to
see if that makes the buster tests more reliable.
This probably isn't part of the tast API so subject to change anytime,
this is fine since we don't really care about execution order, we just
want the tests to run after buster for the next few days.
Doesn't change the display name of the tests.

BUG=chromium:1026021
TEST=tast list and check the order is as expected

Change-Id: I5e20375db07d24642bf1a54d40c663245cbc672f