test-tests: crostini: Don't reuse VM images across runs
Currently if a new version of guest_images.tar is used for a crostini
test it will be ignored if there is already a version of the VM
extracted on disk. This leads to tests mysteriously continuing to
pass/fail even after changes have been made that should effect them.

Fix this by clearing out the directory used to cache these files
instead. This will makes restarting the precondition slightly slower
in some cases, but not very significantly.

Also add some logging so it's easier to tell what version of the VM is running.

BUG=chromium:1028530
TEST=manually tested

Change-Id: Ib4399bf247dc03933fc364d0fbd220c47febd815
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1935953
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>