tast: introduce cache mechanism for PIDs in browser watcher.
It seems that the browser watcher keeps scanning all process
information for every 100msecs, which adds some loads. This
matters more especially on lower-end devices.

This CL introduces caches for the process information, so that
it does not have to scan all of the data everytime.

Theoretically, it is possible that a process finishes and
a new process is created and the same PID is assigned, but
considering the interval of the scan (100msecs), I believe
it is extremely unlikely to happen.

Due to the risk for the misuse of cache, the new cache struct
is private and can only be used by the browwer watcher. Other
users need to scan with an empty cache.

BUG=chromium:1028415
TEST=tast run 'ui.*Perf*', check the CPU load with top

Change-Id: Ib5e14020df7cc7326b9d6b74c313ad360a6f3bb1