tast/webrtc: parameterize webrtc.DecodePerf and add h264 case
ToT webrtc.DecodePerf leaves undetermined the codec profile to
be used; WebRTC will use whichever it feels best for the job,
usually one with hardware enco/decoding support; this means
VP8 usually, but BYT and similarly old devices would pick H264.
This CL parameterises the test case and forces it to be e.g.
h264 since it's the most pervasive (others coming), wiring it
down to the decodePerf() function like its sibling
RunPeerConnection() [1] does.

This also fixes an issue inadvertently introduced by my CL
crrev.com/c/1898721, where the logic changed: before the CL
the hw decode was disabled, and after it was forced-enabled
and expected; before the CL if there was no support for Hw
decoding, we'd just measure twice the sw one, whereas after
the CL, we'll try with both hw and sw, and skip the measure
if the expectation is not met. This CL makes sure we do not
run twice, but also we don't fail if hw decoding is not
supported.

[1] https://chromium.git.corp.google.com/chromiumos/platform/tast-tests/+/1a9cde86fd71ad0a8074674fbd82008069e9deac/src/chromiumos/tast/local/bundles/cros/webrtc/peerconnection/peer_connection.go#34

BUG=chromium:1028414
TEST=tast -verbose run papaya webrtc.DecodePerf.* on candy (BYT)

Change-Id: Icfb3ab88e134bd0c4dc4819eacf18ca0f51da2ad