tast/webrtc: parameterize webrtc.DecodePerf and add h264 case
ToT webrtc.DecodePerf leaves undetermined the codec profile to
be used; WebRTC will use whichever it feels best for the job,
usually one with hardware enco/decoding support, or VP8. This
CL forces it to be e.g. h264 since it's the most pervasive,
wiring it down to the decodePerf() function like its sibling
RunPeerConnection() [1] does, so measurements are comparable.

This also fixes an issue inadvertently introduced by my CL
crrev.com/c/1898721, where the logic changed: before the CL
the hw decode was disabled, and after it was forced-enabled
and expected; before the CL if there was no support for Hw
decoding, we'd just measure twice the sw one, whereas after
the CL, we'd fail (which I think is correct).

[1] https://cs.corp.google.com/chromeos_public/src/platform/tast-tests/src/chromiumos/tast/local/bundles/cros/webrtc/peerconnection/peer_connection.go?type=cs&q=RunPeerConnection+file:%5Esrc/platform/tast-tests/src/chromiumos/tast/local/bundles/cros/+package:%5Echromeos_public$&g=0&l=34

BUG=chromium:1028414
TEST=tast -verbose run papaya webrtc.DecodePerf.* on candy (BYT)

Change-Id: Icfb3ab88e134bd0c4dc4819eacf18ca0f51da2ad