tast-tests: let browser watcher to watch the browser process only
The current code of check method makes on load by getting the
browser process ID for every 100 msecs. It turns out that this scans
all of the process information to find out the browser process. This
requires many CPU cycles and causes some amount of load on lower-end
devices.

Indeed, once it gets the browser process, further tasks would be
watching that process, and there's no special need to scan all
processes everytime.

BUG=chromium:1028415
TEST=tast run 'ui.*Perf*', check the CPU load with top

Change-Id: Ib5e14020df7cc7326b9d6b74c313ad360a6f3bb1